fix(node dependencies): Use package.json for Node deps.
Node.JS dependencies should be defined in the package.json.
This ensures not only that the latest versions of dependencies are found
when running automation tests, but also means that users have less to pull
from the VCS when developing.

You can also use semantic versioning to explicitly define versions to use.